+++ src/main/main.window.ts
@@ -14,7 +14,6 @@ import { autoUpdater } from 'electron-updater';
 import log from 'electron-log';
 import MenuBuilder from './menu';
 import { resolveHtmlPath } from './util';
-import { startBackend } from '../backend/backend';
 
 class AppUpdater {
   constructor() {