+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.1.0] - 2023-10-14
+
+## Add
+- Columna importe en el ticket
+- Consumidor Final como cabecera del ticket.
+- Muchas gracias en el footer del ticket.
+- Boton para limpiar pantalla
+- Eliminacion de menus y cambio de nombre a sanita ticket.
+
 ## [0.0.1] - 2023-10-01
 
 ## Add